+++ build-logic/src/main/kotlin/via.shadow-conventions.gradle.kts
@@ -45,6 +45,7 @@ fun ShadowJar.configureRelocations() {
     relocate("com.google.gson", "com.viaversion.viaversion.libs.gson")
     relocate("it.unimi.dsi.fastutil", "com.viaversion.viaversion.libs.fastutil")
     relocate("net.lenni0451.mcstructs", "com.viaversion.viaversion.libs.mcstructs")
+    relocate("org.yaml.snakeyaml", "com.viaversion.viaversion.libs.snakeyaml")
 }
 
 fun ShadowJar.configureExcludes() {
