--- compat/build.gradle.kts
@@ -1,5 +1,0 @@
-dependencies {
-    api(projects.compat.snakeyamlCompatCommon)
-    api(projects.compat.snakeyaml2Compat)
-    api(projects.compat.snakeyaml1Compat)
-}
