--- compat/snakeyaml-compat-common/build.gradle.kts
@@ -1,3 +1,0 @@
-dependencies {
-    compileOnly(rootProject.libs.snakeYaml2)
-}
