--- compat/snakeyaml2-compat/build.gradle.kts
@@ -1,4 +1,0 @@
-dependencies {
-    api(projects.compat.snakeyamlCompatCommon)
-    compileOnly(rootProject.libs.snakeYaml2)
-}
