+++ gradle/libs.versions.toml
@@ -10,8 +10,7 @@ mcstructs = "5-2.5.0-SNAPSHOT"
 # Common provided
 netty = "4.0.20.Final"
 guava = "17.0"
-snakeYaml = "1.18"
-snakeYaml2 = "2.0"
+snakeYaml = "2.2"
 
 junit = "5.10.2"
 checkerQual = "3.43.0"
@@ -33,7 +32,6 @@ text = { group = "com.viaversion.mcstructs", name = "text", version.ref = "mcstr
 netty = { group = "io.netty", name = "netty-all", version.ref = "netty" }
 guava = { group = "com.google.guava", name = "guava", version.ref = "guava" }
 snakeYaml = { group = "org.yaml", name = "snakeyaml", version.ref = "snakeYaml" }
-snakeYaml2 = { group = "org.yaml", name = "snakeyaml", version.ref = "snakeYaml2" }
 
 jupiterApi = { group = "org.junit.jupiter", name = "junit-jupiter-api", version.ref = "junit" }
 jupiterEngine = { group = "org.junit.jupiter", name = "junit-jupiter-engine", version.ref = "junit" }
